+++ language-support/java/bindings-rxjava/src/main/java/com/daml/ledger/rxjava/CommandClient.java
@@ -4,6 +4,8 @@
 package com.daml.ledger.rxjava;
 
 import com.daml.ledger.javaapi.data.Command;
+import com.daml.ledger.javaapi.data.Transaction;
+import com.daml.ledger.javaapi.data.TransactionTree;
 import com.google.protobuf.Empty;
 import io.reactivex.Single;
 import org.checkerframework.checker.nullness.qual.NonNull;
@@ -19,4 +21,17 @@ public interface CommandClient {
     Single<Empty> submitAndWait(@NonNull String workflowId, @NonNull String applicationId,
                                 @NonNull String commandId, @NonNull String party, @NonNull Instant ledgerEffectiveTime,
                                 @NonNull Instant maximumRecordTime, @NonNull List<@NonNull Command> commands);
+
+    Single<String> submitAndWaitForTransactionId(@NonNull String workflowId, @NonNull String applicationId,
+                                                 @NonNull String commandId, @NonNull String party, @NonNull Instant ledgerEffectiveTime,
+                                                 @NonNull Instant maximumRecordTime, @NonNull List<@NonNull Command> commands);
+
+    Single<Transaction> submitAndWaitForTransaction(@NonNull String workflowId, @NonNull String applicationId,
+                                                    @NonNull String commandId, @NonNull String party, @NonNull Instant ledgerEffectiveTime,
+                                                    @NonNull Instant maximumRecordTime, @NonNull List<@NonNull Command> commands);
+
+    Single<TransactionTree> submitAndWaitForTransactionTree(@NonNull String workflowId, @NonNull String applicationId,
+                                                            @NonNull String commandId, @NonNull String party, @NonNull Instant ledgerEffectiveTime,
+                                                            @NonNull Instant maximumRecordTime, @NonNull List<@NonNull Command> commands);
+
 }