+++ language-support/java/testkit/src/main/scala/com/daml/ledger/testkit/services/CommandServiceImpl.scala
@@ -4,19 +4,42 @@
 package com.daml.ledger.testkit.services
 
 import com.digitalasset.ledger.api.v1.command_service.CommandServiceGrpc.CommandService
-import com.digitalasset.ledger.api.v1.command_service.{CommandServiceGrpc, SubmitAndWaitRequest}
+import com.digitalasset.ledger.api.v1.command_service._
 import com.google.protobuf.empty.Empty
 import io.grpc.ServerServiceDefinition
 
 import scala.concurrent.{ExecutionContext, Future}
 
-class CommandServiceImpl(response: Future[Empty]) extends CommandService {
+class CommandServiceImpl(
+    submitAndWaitResponse: Future[Empty],
+    submitAndWaitForTransactionIdResponse: Future[SubmitAndWaitForTransactionIdResponse],
+    submitAndWaitForTransactionResponse: Future[SubmitAndWaitForTransactionResponse],
+    submitAndWaitForTransactionTreeResponse: Future[SubmitAndWaitForTransactionTreeResponse])
+    extends CommandService {
 
   private var lastRequest: Option[SubmitAndWaitRequest] = None
 
   override def submitAndWait(request: SubmitAndWaitRequest): Future[Empty] = {
     this.lastRequest = Some(request)
-    response
+    submitAndWaitResponse
+  }
+
+  override def submitAndWaitForTransactionId(
+      request: SubmitAndWaitRequest): Future[SubmitAndWaitForTransactionIdResponse] = {
+    this.lastRequest = Some(request)
+    submitAndWaitForTransactionIdResponse
+  }
+
+  override def submitAndWaitForTransaction(
+      request: SubmitAndWaitRequest): Future[SubmitAndWaitForTransactionResponse] = {
+    this.lastRequest = Some(request)
+    submitAndWaitForTransactionResponse
+  }
+
+  override def submitAndWaitForTransactionTree(
+      request: SubmitAndWaitRequest): Future[SubmitAndWaitForTransactionTreeResponse] = {
+    this.lastRequest = Some(request)
+    submitAndWaitForTransactionTreeResponse
   }
 
   def getLastRequest: Option[SubmitAndWaitRequest] = this.lastRequest
@@ -24,9 +47,17 @@ class CommandServiceImpl(response: Future[Empty]) extends CommandService {
 
 object CommandServiceImpl {
 
-  def createWithRef(response: Future[Empty])(
+  def createWithRef(
+      submitAndWaitResponse: Future[Empty],
+      submitAndWaitForTransactionIdResponse: Future[SubmitAndWaitForTransactionIdResponse],
+      submitAndWaitForTransactionResponse: Future[SubmitAndWaitForTransactionResponse],
+      submitAndWaitForTransactionTreeResponse: Future[SubmitAndWaitForTransactionTreeResponse])(
       implicit ec: ExecutionContext): (ServerServiceDefinition, CommandServiceImpl) = {
-    val serviceImpl = new CommandServiceImpl(response)
+    val serviceImpl = new CommandServiceImpl(
+      submitAndWaitResponse,
+      submitAndWaitForTransactionIdResponse,
+      submitAndWaitForTransactionResponse,
+      submitAndWaitForTransactionTreeResponse)
     (CommandServiceGrpc.bindService(serviceImpl, ec), serviceImpl)
   }
 }