+++ ledger/ledger-api-common/src/main/scala/com/digitalasset/platform/server/api/validation/CommandServiceValidation.scala
@@ -4,7 +4,7 @@
 package com.digitalasset.platform.server.api.validation
 
 import com.digitalasset.ledger.api.v1.command_service.CommandServiceGrpc.CommandService
-import com.digitalasset.ledger.api.v1.command_service.{CommandServiceGrpc, SubmitAndWaitRequest}
+import com.digitalasset.ledger.api.v1.command_service._
 import com.digitalasset.platform.api.grpc.GrpcApiService
 import com.digitalasset.platform.common.util.DirectExecutionContext
 import com.digitalasset.platform.server.api.ProxyCloseable
@@ -29,6 +29,24 @@ class CommandServiceValidation(
   override def submitAndWait(request: SubmitAndWaitRequest): Future[Empty] = {
     val validation = requirePresence(request.commands, "commands").flatMap(validateCommands)
     validation.fold(Future.failed, _ => service.submitAndWait(request))
+  }
+
+  override def submitAndWaitForTransactionId(
+      request: SubmitAndWaitRequest): Future[SubmitAndWaitForTransactionIdResponse] = {
+    val validation = requirePresence(request.commands, "commands").flatMap(validateCommands)
+    validation.fold(Future.failed, _ => service.submitAndWaitForTransactionId(request))
+  }
+
+  override def submitAndWaitForTransaction(
+      request: SubmitAndWaitRequest): Future[SubmitAndWaitForTransactionResponse] = {
+    val validation = requirePresence(request.commands, "commands").flatMap(validateCommands)
+    validation.fold(Future.failed, _ => service.submitAndWaitForTransaction(request))
+  }
+
+  override def submitAndWaitForTransactionTree(
+      request: SubmitAndWaitRequest): Future[SubmitAndWaitForTransactionTreeResponse] = {
+    val validation = requirePresence(request.commands, "commands").flatMap(validateCommands)
+    validation.fold(Future.failed, _ => service.submitAndWaitForTransactionTree(request))
   }
 
   override def bindService(): ServerServiceDefinition =