+++ ledger/ledger-api-common/src/main/scala/com/digitalasset/platform/server/services/command/ReferenceCommandService.scala
@@ -16,7 +16,14 @@ import com.digitalasset.ledger.api.v1.command_completion_service.{
 }
 import com.digitalasset.ledger.api.v1.command_service._
 import com.digitalasset.ledger.api.v1.command_submission_service.CommandSubmissionServiceGrpc.CommandSubmissionService
-import com.digitalasset.ledger.api.v1.command_submission_service.{SubmitRequest}
+import com.digitalasset.ledger.api.v1.command_submission_service.SubmitRequest
+import com.digitalasset.ledger.api.v1.completion.Completion
+import com.digitalasset.ledger.api.v1.transaction_service.TransactionServiceGrpc.TransactionService
+import com.digitalasset.ledger.api.v1.transaction_service.{
+  GetFlatTransactionResponse,
+  GetTransactionByIdRequest,
+  GetTransactionResponse
+}
 import com.digitalasset.ledger.client.configuration.CommandClientConfiguration
 import com.digitalasset.ledger.client.services.commands.{
   CommandClient,
@@ -85,61 +92,113 @@ class ReferenceCommandService private (
     )
 
   @SuppressWarnings(Array("org.wartremover.warts.Any"))
-  override def submitAndWait(request: SubmitAndWaitRequest): Future[Empty] = {
+  private def submitAndWaitInternal(request: SubmitAndWaitRequest): Future[Completion] = {
 
     val appId = request.getCommands.applicationId
     val submitter = TrackerMap.Key(application = appId, party = request.getCommands.party)
 
     if (running) {
-      submissionTracker.track(submitter, request) {
-        for {
-          trackingFlow <- {
-            lowLevelCommandServiceAccess match {
-              case LowLevelCommandServiceAccess.RemoteServices(submission, completion) =>
-                val client = commandClient(appId, submission, completion)
-                if (configuration.limitMaxCommandsInFlight)
-                  client.trackCommands[Promise[Empty]](List(submitter.party))
-                else
-                  client.trackCommandsUnbounded[Promise[Empty]](List(submitter.party))
-              case LowLevelCommandServiceAccess.LocalServices(
-                  submissionFlow,
-                  getCompletionSource,
-                  getCompletionEnd) =>
-                for {
-                  ledgerEnd <- getCompletionEnd().map(_.getOffset)
-                } yield {
-                  val tracker =
-                    CommandTrackerFlow[Promise[Empty], NotUsed](
-                      submissionFlow,
-                      offset =>
-                        getCompletionSource(
-                          CompletionStreamRequest(
-                            configuration.ledgerId,
-                            appId,
-                            List(submitter.party),
-                            Some(offset)))
-                          .mapConcat(CommandCompletionSource.toStreamElements),
-                      ledgerEnd
-                    )
-
+      submissionTracker
+        .track(submitter, request) {
+          for {
+            trackingFlow <- {
+              lowLevelCommandServiceAccess match {
+                case LowLevelCommandServiceAccess.RemoteServices(submission, completion, _) =>
+                  val client = commandClient(appId, submission, completion)
                   if (configuration.limitMaxCommandsInFlight)
-                    MaxInFlight(configuration.maxCommandsInFlight).joinMat(tracker)(Keep.right)
-                  else tracker
-                }
+                    client.trackCommands[Promise[Completion]](List(submitter.party))
+                  else
+                    client.trackCommandsUnbounded[Promise[Completion]](List(submitter.party))
+                case LowLevelCommandServiceAccess.LocalServices(
+                    submissionFlow,
+                    getCompletionSource,
+                    getCompletionEnd,
+                    _,
+                    _) =>
+                  for {
+                    ledgerEnd <- getCompletionEnd().map(_.getOffset)
+                  } yield {
+                    val tracker =
+                      CommandTrackerFlow[Promise[Completion], NotUsed](
+                        submissionFlow,
+                        offset =>
+                          getCompletionSource(
+                            CompletionStreamRequest(
+                              configuration.ledgerId,
+                              appId,
+                              List(submitter.party),
+                              Some(offset)))
+                            .mapConcat(CommandCompletionSource.toStreamElements),
+                        ledgerEnd
+                      )
+
+                    if (configuration.limitMaxCommandsInFlight)
+                      MaxInFlight(configuration.maxCommandsInFlight).joinMat(tracker)(Keep.right)
+                    else tracker
+                  }
+              }
             }
+          } yield {
+            TrackerImpl(trackingFlow, configuration.inputBufferSize, configuration.historySize)
           }
-        } yield {
-          TrackerImpl(trackingFlow, configuration.inputBufferSize, configuration.historySize)
         }
-      }
     } else {
       Future.failed(
         new ApiException(Status.UNAVAILABLE.withDescription("Service has been shut down.")))
     }
   }
 
+  override def submitAndWait(request: SubmitAndWaitRequest): Future[Empty] = {
+    submitAndWaitInternal(request).map(_ => Empty.defaultInstance)
+  }
+
+  override def submitAndWaitForTransactionId(
+      request: SubmitAndWaitRequest): Future[SubmitAndWaitForTransactionIdResponse] = {
+    submitAndWaitInternal(request).map { compl =>
+      SubmitAndWaitForTransactionIdResponse(compl.transactionId)
+    }
+  }
+
+  override def submitAndWaitForTransaction(
+      request: SubmitAndWaitRequest): Future[SubmitAndWaitForTransactionResponse] = {
+    submitAndWaitInternal(request).flatMap { resp =>
+      val txRequest = GetTransactionByIdRequest(
+        request.getCommands.ledgerId,
+        resp.transactionId,
+        List(request.getCommands.party))
+      flatById(txRequest).map(resp => SubmitAndWaitForTransactionResponse(resp.transaction))
+
+    }
+  }
+
+  override def submitAndWaitForTransactionTree(
+      request: SubmitAndWaitRequest): Future[SubmitAndWaitForTransactionTreeResponse] = {
+    submitAndWaitInternal(request).flatMap { resp =>
+      val txRequest = GetTransactionByIdRequest(
+        request.getCommands.ledgerId,
+        resp.transactionId,
+        List(request.getCommands.party))
+
+      treeById(txRequest).map(resp => SubmitAndWaitForTransactionTreeResponse(resp.transaction))
+    }
+  }
+
   override def toString: String = ReferenceCommandService.getClass.getSimpleName
 
+  private val (treeById, flatById) = {
+    lowLevelCommandServiceAccess match {
+      case LowLevelCommandServiceAccess.RemoteServices(_, _, transaction) =>
+        (transaction.getTransactionById _, transaction.getFlatTransactionById _)
+
+      case LowLevelCommandServiceAccess.LocalServices(
+          _,
+          _,
+          _,
+          getTransactionById,
+          getFlatTransactionById) =>
+        (getTransactionById, getFlatTransactionById)
+    }
+  }
 }
 
 object ReferenceCommandService {
@@ -170,16 +229,19 @@ object ReferenceCommandService {
 
     final case class RemoteServices(
         submissionStub: CommandSubmissionService,
-        completionStub: CommandCompletionService)
+        completionStub: CommandCompletionService,
+        transactionService: TransactionService)
         extends LowLevelCommandServiceAccess
 
     final case class LocalServices(
         submissionFlow: Flow[
-          Ctx[(Promise[Empty], String), SubmitRequest],
-          Ctx[(Promise[Empty], String), Try[Empty]],
+          Ctx[(Promise[Completion], String), SubmitRequest],
+          Ctx[(Promise[Completion], String), Try[Empty]],
           NotUsed],
         getCompletionSource: CompletionStreamRequest => Source[CompletionStreamResponse, NotUsed],
-        getCompletionEnd: () => Future[CompletionEndResponse])
+        getCompletionEnd: () => Future[CompletionEndResponse],
+        getTransactionById: GetTransactionByIdRequest => Future[GetTransactionResponse],
+        getFlatTransactionById: GetTransactionByIdRequest => Future[GetFlatTransactionResponse])
         extends LowLevelCommandServiceAccess
 
   }