+++ ledger/ledger-api-integration-tests/src/test/lib/scala/com/digitalasset/platform/tests/integration/ledger/api/TransactionServiceHelpers.scala
@@ -69,8 +69,9 @@ trait TransactionServiceHelpers extends Matchers {
     } yield c)
       .via(trackingFlow)
       .runWith(Sink.foreach {
-        case Ctx(i, Completion(_, Some(status), _)) =>
+        case Ctx(i, Completion(_, Some(status), transactionId, _)) =>
           status should have('code (0))
+          transactionId should not be empty
           ()
       })
   }