+++ ledger/sandbox/src/main/scala/com/digitalasset/ledger/server/LedgerApiServer/LedgerApiServer.scala
@@ -129,7 +129,9 @@ object LedgerApiServer {
           completionService.service
             .asInstanceOf[SandboxCommandCompletionService]
             .completionStreamSource(r),
-        () => completionService.completionEnd(CompletionEndRequest(ledgerBackend.ledgerId))
+        () => completionService.completionEnd(CompletionEndRequest(ledgerBackend.ledgerId)),
+        transactionService.getTransactionById,
+        transactionService.getFlatTransactionById
       )
     )